reworked configuration
