+++ src/builder/DruidQueryBuilder.tsx
@@ -4,7 +4,7 @@ import { QueryBuilderProps } from './types';
 import { debounce } from 'lodash';
 
 export const DruidQueryBuilder = (props: QueryBuilderProps) => {
-  const debouncedOnOptionsChange = debounce(props.onOptionsChange, props.options.settings.debounceTime || 250);
-  const newProps = { ...props, onOptionsChange: debouncedOnOptionsChange };
-  return <Query {...newProps} />;
+  return (
+    <Query {...props} onOptionsChange={debounce(props.onOptionsChange, props.options.settings.debounceTime || 250)} />
+  );
 };
