+++ src/configuration/QuerySettings/DruidQuerySettings.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { FieldSet } from '@grafana/ui';
 import { css, cx } from 'emotion';
-import { DruidQueryContextSettings, DruidQueryResponseSettings } from './';
+import { DruidQueryRequestSettings, DruidQueryResponseSettings } from './';
 import { QuerySettingsProps } from './types';
 
 export const DruidQuerySettings = (props: QuerySettingsProps) => {
   return (
     <>
-      <FieldSet label="Query" className={cx(styles.fieldset)}>
-        <DruidQueryContextSettings {...props} />
+      <FieldSet label="Request" className={cx(styles.fieldset)}>
+        <DruidQueryRequestSettings {...props} />
       </FieldSet>
       <FieldSet label="Response" className={cx(styles.fieldset)}>
         <DruidQueryResponseSettings {...props} />
