+++ src/configuration/QuerySettings/index.ts
@@ -1,5 +1,6 @@
 export { DruidQueryDefaultSettings } from './DruidQueryDefaultSettings';
 export { DruidQuerySettings } from './DruidQuerySettings';
+export { DruidQueryRequestSettings } from './DruidQueryRequestSettings';
 export { DruidQueryContextSettings } from './DruidQueryContextSettings';
 export { DruidQueryResponseSettings } from './DruidQueryResponseSettings';
 export { DruidQueryLogSettings } from './DruidQueryLogSettings';
